requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
beanie==1.20.0
cffi==1.15.1
click==8.1.4
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.0
h11==0.14.0
idna==3.4
lazy-model==0.0.5
motor==3.2.0
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.1.2
pymongo==4.4.0
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.7.1
uvicorn==0.22.0
